trip_admin/requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.7
django-admin-rangefilter==0.11.2
django-extensions==3.2.3
django-postgres-extra==2.0.8
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.13.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
gunicorn==21.2.0
identify==2.5.32
idna==3.4
inflect==7.0.0
magic-filter==1.0.12
makefile==1.1.0
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.0.0
pre-commit==3.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
scheduler==0.8.5
six==1.16.0
smmap==5.0.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.50
sqlparse==0.4.4
typeguard==4.1.5
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
virtualenv==20.24.7
wrapt==1.16.0
yarl==1.9.2
